requirements.txt
paramiko
liboqs
cryptography @ file:///home/juanito/Documents/UPM/TFG/Cryptography/cryptography-43.0.3